refactor: rename parse file
