--- backend/apps/utils/parse_bool.py
@@ -1,12 +1,0 @@
-from typing import TypeVar
-
-T = TypeVar("T")
-
-
-def parse_bool(value: str | None, default: T = None) -> bool | T:
-    if value is None:
-        return default
-
-    formatted_value = value.lower().replace("/", "")
-
-    return formatted_value not in ("false", "0")
